requirements.txt
Flask==2.1.4
flask-jwt-extended==4.2.0
flasgger==0.9.5
mysql-connector-python==8.0.28
requests==2.26.0
